Use float values for height

+++ mvtjson2oscimv4.py
@@ -527,7 +527,7 @@ def convert(tile_z,tile_x,tile_y,buffer_pixels,fr):
                     fixed_kv['building'] = 'yes'
 
                     if 'height' in kv:
-                        fixed_kv['height'] = str(int(kv['height'])*100)
+                        fixed_kv['height'] = str(float(kv['height'])*100.0)
 
                 elif type_ in frozenset([
                         'bar',
